requirements.txt
Django==2.0.2
psycopg2==2.7.4
djangorestframework==3.8.2
django-webpack-loader==0.6.0
social-auth-app-django
Pillow==5.1.0
